# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant ERP \u0026 Commerce SaaS API",
        "title": "CRM System Sales API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate user and return JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/invoice": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new draft invoice for a buyer and seller relationship",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice"
                ],
                "summary": "Create invoice draft",
                "parameters": [
                    {
                        "description": "Invoice data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invoicedto.CreateInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.CreateInvoiceSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/invoice/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns invoice details and financial information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice"
                ],
                "summary": "Get invoice by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.GetInvoiceSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/invoice/{id}/pay": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers a payment for an invoice and updates payment status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice"
                ],
                "summary": "Pay invoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invoicedto.PayInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.PayInvoiceSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/invoice/{id}/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates invoice items, reserves stock and changes invoice status from draft to pending",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice"
                ],
                "summary": "Submit invoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SubmitInvoiceSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "authdto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "docs.CreateInvoiceSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/invoicedto.InvoiceResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "docs.GetInvoiceSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/invoicedto.GetInvoiceResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "docs.PayInvoiceSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/invoicedto.PayInvoiceResponse"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "docs.SubmitInvoiceSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "string",
                    "example": "invoice enviada correctamente"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "invoicedto.CreateInvoiceRequest": {
            "type": "object",
            "required": [
                "buyer_client_id",
                "seller_company_id"
            ],
            "properties": {
                "buyer_client_id": {
                    "type": "integer",
                    "example": 5
                },
                "seller_company_id": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "invoicedto.GetInvoiceResponse": {
            "type": "object",
            "properties": {
                "buyer_client_id": {
                    "type": "integer",
                    "example": 8
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-06-15T10:00:00Z"
                },
                "created_by_user_id": {
                    "type": "integer",
                    "example": 12
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "paid_amount": {
                    "type": "number",
                    "example": 500
                },
                "seller_company_id": {
                    "type": "integer",
                    "example": 2
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "status_invoice": {
                    "type": "string",
                    "example": "pending"
                },
                "subtotal": {
                    "type": "number",
                    "example": 1000
                },
                "taxes": {
                    "type": "number",
                    "example": 210
                },
                "total_amount": {
                    "type": "number",
                    "example": 1210
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "invoicedto.InvoiceResponse": {
            "type": "object",
            "properties": {
                "buyer_client_id": {
                    "type": "integer",
                    "example": 5
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-06-15T18:30:00Z"
                },
                "created_by_user_id": {
                    "type": "integer",
                    "example": 10
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "seller_company_id": {
                    "type": "integer",
                    "example": 2
                },
                "status_invoice": {
                    "type": "string",
                    "example": "draft"
                },
                "total_amount": {
                    "type": "number",
                    "example": 0
                }
            }
        },
        "invoicedto.PayInvoiceRequest": {
            "type": "object",
            "required": [
                "amount",
                "payment_method"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 1500
                },
                "payment_method": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "cash"
                }
            }
        },
        "invoicedto.PayInvoiceResponse": {
            "type": "object",
            "properties": {
                "paid_amount": {
                    "type": "number",
                    "example": 1500
                },
                "payment_id": {
                    "type": "integer",
                    "example": 21
                },
                "remaining_amount": {
                    "type": "number",
                    "example": 0
                },
                "status_invoice": {
                    "type": "string",
                    "example": "paid"
                },
                "total_amount": {
                    "type": "number",
                    "example": 1500
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "errorMessage": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}